creceEcuador/requirements.txt
appdirs==1.4.3
asgiref==3.2.7
attrs==18.2.0
boto3==1.4.4
botocore==1.5.62
cached-property==1.4.3
cachetools==2.1.0
certifi==2018.8.24
chardet==3.0.4
click==6.6
defusedxml==0.5.0
Django==3.0.5
djangorestframework==3.11.0
docutils==0.12
filechunkio==1.8
futures==3.1.1
google-auth==1.5.1
google-auth-oauthlib==0.2.0
googleads==13.0.0
idna==2.7
isodate==0.6.0
jmespath==0.9.0
lxml==4.2.4
numpy==1.15.1
oauthlib==2.1.0
pandas==0.23.4
pbr==4.2.0
Pillow==7.0.0
pyaml==15.8.2
pyasn1==0.4.4
pyasn1-modules==0.2.2
python-dateutil==2.5.3
python-lambda==3.2.2
pytz==2018.5
PyYAML==3.11
requests==2.19.1
requests-oauthlib==1.0.0
requests-toolbelt==0.8.0
rsa==3.4.2
s3transfer==0.1.13
six==1.10.0
sqlparse==0.3.1
stevedore==1.29.0
suds-jurko==0.6
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
xmltodict==0.11.0
zeep==3.1.0
